refactor: MovingStrategy 구현체 생성 부분 팩토리 메서드로 분리, strategy 패키지 분리
