+++ src/main/java/racingcar/Cars.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.strategy.MovingStrategy;
+
 import java.util.ArrayList;
 import java.util.List;
