+++ src/main/java/racingcar/strategy/MovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.strategy;
 
 @FunctionalInterface
 public interface MovingStrategy {
