+++ src/main/java/racingcar/strategy/MovingStrategyType.java
@@ -1,0 +1,5 @@
+package racingcar.strategy;
+
+public enum MovingStrategyType {
+    RANDOM
+}
